Larger refactor. Too many query methods in setup. Next cleanup PR will organize this better.
