+++ test/unit/php/includes/tests/core/classes/class-test-event-query.php
@@ -41,10 +41,46 @@ class Test_Event_Query extends Base {
 				'callback' => array( $instance, 'prepare_event_query_before_execution' ),
 			),
 			array(
+				'type'     => 'action',
+				'name'     => 'pre_get_posts',
+				'priority' => 10,
+				'callback' => array( $instance, 'handle_rsvp_sorting' ),
+			),
+			array(
+				'type'     => 'action',
+				'name'     => 'pre_get_posts',
+				'priority' => 10,
+				'callback' => array( $instance, 'handle_venue_sorting' ),
+			),
+			array(
 				'type'     => 'filter',
 				'name'     => 'posts_clauses',
 				'priority' => 9,
 				'callback' => array( $instance, 'adjust_admin_event_sorting' ),
+			),
+			array(
+				'type'     => 'action',
+				'name'     => 'load-edit.php',
+				'priority' => 10,
+				'callback' => array( $instance, 'default_sort' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => sprintf( 'manage_edit-%s_sortable_columns', Event::POST_TYPE ),
+				'priority' => 10,
+				'callback' => array( $instance, 'sortable_columns' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => sprintf( 'views_edit-%s', Event::POST_TYPE ),
+				'priority' => 10,
+				'callback' => array( $instance, 'views_edit' ),
+			),
+			array(
+				'type'     => 'filter',
+				'name'     => 'query_vars',
+				'priority' => 10,
+				'callback' => array( $instance, 'query_vars' ),
 			),
 		);
 
@@ -1415,6 +1451,1274 @@ class Test_Event_Query extends Base {
 			$future_post->ID,
 			$past->posts,
 			'Future event should not appear in past events query'
+		);
+	}
+
+	/**
+	 * Coverage for handle_rsvp_sorting method.
+	 *
+	 * @covers ::handle_rsvp_sorting
+	 *
+	 * @return void
+	 */
+	public function test_handle_rsvp_sorting(): void {
+		$instance = Event_Query::get_instance();
+
+		// Create a mock query.
+		$query = $this->createMock( \WP_Query::class );
+
+		// Test non-admin context (is_admin() returns false by default in tests).
+		$query->method( 'is_main_query' )->willReturn( true );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, Event::POST_TYPE ),
+				array( 'orderby', null, 'rsvps' ),
+			)
+		);
+
+		// Should return early due to non-admin context.
+		$instance->handle_rsvp_sorting( $query );
+
+		// Test different post type - should return early.
+		$query = $this->createMock( \WP_Query::class );
+		$query->method( 'is_main_query' )->willReturn( true );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, 'post' ),
+				array( 'orderby', null, 'rsvps' ),
+			)
+		);
+
+		$instance->handle_rsvp_sorting( $query );
+
+		// Test non-main query - should return early.
+		$query = $this->createMock( \WP_Query::class );
+		$query->method( 'is_main_query' )->willReturn( false );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, Event::POST_TYPE ),
+				array( 'orderby', null, 'rsvps' ),
+			)
+		);
+
+		$instance->handle_rsvp_sorting( $query );
+
+		// Test different orderby - should return early.
+		$query = $this->createMock( \WP_Query::class );
+		$query->method( 'is_main_query' )->willReturn( true );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, Event::POST_TYPE ),
+				array( 'orderby', null, 'title' ),
+			)
+		);
+
+		$instance->handle_rsvp_sorting( $query );
+
+		// Since the method primarily adds filters and sets query vars,
+		// and we can't easily test is_admin() in unit tests,
+		// we'll focus on testing the individual components.
+		$this->assertTrue( true ); // Method completed without errors.
+	}
+
+	/**
+	 * Coverage for rsvp_sorting_join_paged method.
+	 *
+	 * @covers ::rsvp_sorting_join_paged
+	 *
+	 * @return void
+	 */
+	public function test_rsvp_sorting_join_paged(): void {
+		global $wpdb;
+		$instance = Event_Query::get_instance();
+
+		$original_join = "LEFT JOIN {$wpdb->posts} AS posts ON posts.ID = {$wpdb->posts}.ID";
+		$result        = $instance->rsvp_sorting_join_paged( $original_join );
+
+		// Should contain the original join plus the RSVP join.
+		$this->assertStringContainsString( $original_join, $result );
+		$this->assertStringContainsString( 'LEFT JOIN', $result );
+		$this->assertStringContainsString( $wpdb->comments, $result );
+		$this->assertStringContainsString( 'rsvp_sort_comments', $result );
+		$this->assertStringContainsString( 'comment_type', $result );
+		$this->assertStringContainsString( "comment_approved = '1'", $result );
+	}
+
+	/**
+	 * Coverage for sorting_groupby_post_id method.
+	 *
+	 * @covers ::sorting_groupby_post_id
+	 *
+	 * @return void
+	 */
+	public function test_sorting_groupby_post_id(): void {
+		global $wpdb;
+		$instance = Event_Query::get_instance();
+
+		$result = $instance->sorting_groupby_post_id( '' );
+		$this->assertEquals( "`{$wpdb->posts}`.`ID`", $result );
+
+		// Test with existing groupby - should keep the existing value.
+		$existing_groupby = 'existing_group';
+		$result           = $instance->sorting_groupby_post_id( $existing_groupby );
+		$this->assertEquals( 'existing_group', $result );
+	}
+
+	/**
+	 * Coverage for rsvp_sorting_orderby method.
+	 *
+	 * Note: This method relies on the global $wp_query, so we test the method's structure
+	 * and ensure it returns a proper ORDER BY clause with expected patterns.
+	 *
+	 * @covers ::rsvp_sorting_orderby
+	 *
+	 * @return void
+	 */
+	public function test_rsvp_sorting_orderby(): void {
+		$instance = Event_Query::get_instance();
+
+		$result = $instance->rsvp_sorting_orderby( 'original_orderby' );
+
+		// Should contain the expected COUNT structure.
+		$this->assertStringContainsString( 'COUNT(rsvp_sort_comments.comment_ID)', $result );
+
+		// Should contain either ASC or DESC (defaults to ASC).
+		$this->assertTrue(
+			strpos( $result, 'ASC' ) !== false || strpos( $result, 'DESC' ) !== false,
+			'ORDER BY clause should contain ASC or DESC'
+		);
+
+		// Verify the method returns a string (basic type check).
+		$this->assertIsString( $result );
+		$this->assertNotEmpty( $result );
+	}
+
+	/**
+	 * Coverage for handle_venue_sorting method.
+	 *
+	 * @covers ::handle_venue_sorting
+	 *
+	 * @return void
+	 */
+	public function test_handle_venue_sorting(): void {
+		$instance = Event_Query::get_instance();
+
+		// Create a mock query.
+		$query = $this->createMock( \WP_Query::class );
+
+		// Test non-admin context (is_admin() returns false by default in tests).
+		$query->method( 'is_main_query' )->willReturn( true );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, Event::POST_TYPE ),
+				array( 'orderby', null, 'venue' ),
+			)
+		);
+
+		// Should return early due to non-admin context.
+		$instance->handle_venue_sorting( $query );
+
+		// Test different post type - should return early.
+		$query = $this->createMock( \WP_Query::class );
+		$query->method( 'is_main_query' )->willReturn( true );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, 'post' ),
+				array( 'orderby', null, 'venue' ),
+			)
+		);
+
+		$instance->handle_venue_sorting( $query );
+
+		// Test non-main query - should return early.
+		$query = $this->createMock( \WP_Query::class );
+		$query->method( 'is_main_query' )->willReturn( false );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, Event::POST_TYPE ),
+				array( 'orderby', null, 'venue' ),
+			)
+		);
+
+		$instance->handle_venue_sorting( $query );
+
+		// Test different orderby - should return early.
+		$query = $this->createMock( \WP_Query::class );
+		$query->method( 'is_main_query' )->willReturn( true );
+		$query->method( 'get' )->willReturnMap(
+			array(
+				array( 'post_type', null, Event::POST_TYPE ),
+				array( 'orderby', null, 'title' ),
+			)
+		);
+
+		$instance->handle_venue_sorting( $query );
+
+		// Since the method primarily adds filters and sets query vars,
+		// and we can't easily test is_admin() in unit tests,
+		// we'll focus on testing the individual components.
+		$this->assertTrue( true ); // Method completed without errors.
+	}
+
+	/**
+	 * Coverage for venue_sorting_join_paged method.
+	 *
+	 * @covers ::venue_sorting_join_paged
+	 *
+	 * @return void
+	 */
+	public function test_venue_sorting_join_paged(): void {
+		global $wpdb;
+		$instance = Event_Query::get_instance();
+
+		$original_join = "LEFT JOIN {$wpdb->posts} AS posts ON posts.ID = {$wpdb->posts}.ID";
+		$result        = $instance->venue_sorting_join_paged( $original_join );
+
+		// Should contain the original join plus the venue joins.
+		$this->assertStringContainsString( $original_join, $result );
+		$this->assertStringContainsString( 'LEFT JOIN', $result );
+		$this->assertStringContainsString( $wpdb->term_relationships, $result );
+		$this->assertStringContainsString( 'venue_tr', $result );
+		$this->assertStringContainsString( $wpdb->term_taxonomy, $result );
+		$this->assertStringContainsString( 'venue_tt', $result );
+		$this->assertStringContainsString( $wpdb->terms, $result );
+		$this->assertStringContainsString( 'venue_terms', $result );
+		$this->assertStringContainsString( "'" . Venue::TAXONOMY . "'", $result );
+	}
+
+	/**
+	 * Coverage for venue_sorting_orderby method.
+	 *
+	 * Note: This method relies on the global $wp_query, so we test the method's structure
+	 * and ensure it returns a proper ORDER BY clause with expected patterns.
+	 *
+	 * @covers ::venue_sorting_orderby
+	 *
+	 * @return void
+	 */
+	public function test_venue_sorting_orderby(): void {
+		$instance = Event_Query::get_instance();
+
+		$result = $instance->venue_sorting_orderby( 'original_orderby' );
+
+		// Should contain the expected CASE structure for NULL handling.
+		$this->assertStringContainsString( 'CASE WHEN venue_terms.name IS NULL THEN 1 ELSE 0 END ASC', $result );
+
+		// Should contain venue_terms.name in the ORDER BY.
+		$this->assertStringContainsString( 'venue_terms.name', $result );
+
+		// Should contain either ASC or DESC (defaults to ASC).
+		$this->assertTrue(
+			strpos( $result, 'ASC' ) !== false || strpos( $result, 'DESC' ) !== false,
+			'ORDER BY clause should contain ASC or DESC'
+		);
+
+		// Verify the method returns a string (basic type check).
+		$this->assertIsString( $result );
+		$this->assertNotEmpty( $result );
+	}
+
+	/**
+	 * Tests handle_rsvp_sorting with RSVP orderby.
+	 *
+	 * Covers: RSVP sorting logic.
+	 *
+	 * @covers ::handle_rsvp_sorting
+	 * @return void
+	 */
+	public function test_rsvp_sorting_with_rsvp_orderby(): void {
+		global $wp_the_query;
+
+		// Create a WP_Query for events with RSVP sorting.
+		$query = new WP_Query(
+			array(
+				'post_type' => Event::POST_TYPE,
+				'orderby'   => 'rsvps',
+				'order'     => 'DESC',
+			)
+		);
+
+		// Set as main query.
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited -- Necessary for testing query modifications.
+		$wp_the_query = $query;
+
+		// Simulate admin context.
+		set_current_screen( 'edit-gatherpress_event' );
+
+		$instance = Event_Query::get_instance();
+		$instance->handle_rsvp_sorting( $query );
+
+		// Verify that sorting order was set.
+		$this->assertEquals( 'DESC', $query->get( 'rsvp_sort_order' ), 'Should set DESC order' );
+
+		// Verify filters were added.
+		$this->assertNotFalse(
+			has_filter( 'posts_join_paged', array( $instance, 'rsvp_sorting_join_paged' ) ),
+			'Should add posts_join_paged filter'
+		);
+		$this->assertNotFalse(
+			has_filter( 'posts_groupby', array( $instance, 'sorting_groupby_post_id' ) ),
+			'Should add posts_groupby filter'
+		);
+		$this->assertNotFalse(
+			has_filter( 'posts_orderby', array( $instance, 'rsvp_sorting_orderby' ) ),
+			'Should add posts_orderby filter'
+		);
+
+		// Clean up.
+		remove_filter( 'posts_join_paged', array( $instance, 'rsvp_sorting_join_paged' ) );
+		remove_filter( 'posts_groupby', array( $instance, 'sorting_groupby_post_id' ) );
+		remove_filter( 'posts_orderby', array( $instance, 'rsvp_sorting_orderby' ) );
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Tests handle_rsvp_sorting with invalid order.
+	 *
+	 * Covers: Order validation.
+	 *
+	 * @covers ::handle_rsvp_sorting
+	 * @return void
+	 */
+	public function test_rsvp_sorting_with_invalid_order(): void {
+		global $wp_the_query;
+
+		// Create a WP_Query for RSVP sorting.
+		$query = new WP_Query(
+			array(
+				'post_type' => Event::POST_TYPE,
+				'orderby'   => 'rsvps',
+			)
+		);
+
+		// Manually set invalid order (bypasses WP_Query's sanitization).
+		$query->set( 'order', 'INVALID' );
+
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited -- Necessary for testing query modifications.
+		$wp_the_query = $query;
+
+		set_current_screen( 'edit-gatherpress_event' );
+
+		$instance = Event_Query::get_instance();
+		$instance->handle_rsvp_sorting( $query );
+
+		// Should default to ASC for invalid order.
+		$this->assertEquals( 'ASC', $query->get( 'rsvp_sort_order' ), 'Should default to ASC for invalid order' );
+
+		// Clean up.
+		remove_filter( 'posts_join_paged', array( $instance, 'rsvp_sorting_join_paged' ) );
+		remove_filter( 'posts_groupby', array( $instance, 'sorting_groupby_post_id' ) );
+		remove_filter( 'posts_orderby', array( $instance, 'rsvp_sorting_orderby' ) );
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Tests handle_rsvp_sorting early return when orderby is not 'rsvps'.
+	 *
+	 * @covers ::handle_rsvp_sorting
+	 * @return void
+	 */
+	public function test_rsvp_sorting_early_return_wrong_orderby(): void {
+		global $wp_the_query;
+
+		// Create a WP_Query for non-RSVP sorting.
+		$query = new WP_Query(
+			array(
+				'post_type' => Event::POST_TYPE,
+				'orderby'   => 'date',
+			)
+		);
+
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited -- Necessary for testing query modifications.
+		$wp_the_query = $query;
+
+		set_current_screen( 'edit-gatherpress_event' );
+
+		$instance = Event_Query::get_instance();
+		$instance->handle_rsvp_sorting( $query );
+
+		// Should not set rsvp_sort_order since orderby is not 'rsvps'.
+		$this->assertEmpty( $query->get( 'rsvp_sort_order' ), 'Should not set rsvp_sort_order for non-RSVP orderby' );
+
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Tests handle_venue_sorting with venue orderby.
+	 *
+	 * Covers: Venue sorting logic.
+	 *
+	 * @covers ::handle_venue_sorting
+	 * @return void
+	 */
+	public function test_venue_sorting_with_venue_orderby(): void {
+		global $wp_the_query;
+
+		// Create a WP_Query for events with venue sorting.
+		$query = new WP_Query(
+			array(
+				'post_type' => Event::POST_TYPE,
+				'orderby'   => 'venue',
+				'order'     => 'DESC',
+			)
+		);
+
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited -- Necessary for testing query modifications.
+		$wp_the_query = $query;
+
+		set_current_screen( 'edit-gatherpress_event' );
+
+		$instance = Event_Query::get_instance();
+		$instance->handle_venue_sorting( $query );
+
+		// Verify that sorting order was set.
+		$this->assertEquals( 'DESC', $query->get( 'venue_sort_order' ), 'Should set DESC order' );
+
+		// Verify filters were added.
+		$this->assertNotFalse(
+			has_filter( 'posts_join_paged', array( $instance, 'venue_sorting_join_paged' ) ),
+			'Should add posts_join_paged filter for venue sorting'
+		);
+		$this->assertNotFalse(
+			has_filter( 'posts_groupby', array( $instance, 'sorting_groupby_post_id' ) ),
+			'Should add posts_groupby filter for venue sorting'
+		);
+		$this->assertNotFalse(
+			has_filter( 'posts_orderby', array( $instance, 'venue_sorting_orderby' ) ),
+			'Should add posts_orderby filter for venue sorting'
+		);
+
+		// Clean up.
+		remove_filter( 'posts_join_paged', array( $instance, 'venue_sorting_join_paged' ) );
+		remove_filter( 'posts_groupby', array( $instance, 'sorting_groupby_post_id' ) );
+		remove_filter( 'posts_orderby', array( $instance, 'venue_sorting_orderby' ) );
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Tests handle_venue_sorting with invalid order.
+	 *
+	 * Covers: Order validation in venue sorting.
+	 *
+	 * @covers ::handle_venue_sorting
+	 * @return void
+	 */
+	public function test_venue_sorting_with_invalid_order(): void {
+		global $wp_the_query;
+
+		// Create a WP_Query for venue sorting.
+		$query = new WP_Query(
+			array(
+				'post_type' => Event::POST_TYPE,
+				'orderby'   => 'venue',
+			)
+		);
+
+		// Manually set invalid order (bypasses WP_Query's sanitization).
+		$query->set( 'order', 'RANDOM' );
+
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited -- Necessary for testing query modifications.
+		$wp_the_query = $query;
+
+		set_current_screen( 'edit-gatherpress_event' );
+
+		$instance = Event_Query::get_instance();
+		$instance->handle_venue_sorting( $query );
+
+		// Should default to ASC for invalid order.
+		$this->assertEquals( 'ASC', $query->get( 'venue_sort_order' ), 'Should default to ASC for invalid order' );
+
+		// Clean up.
+		remove_filter( 'posts_join_paged', array( $instance, 'venue_sorting_join_paged' ) );
+		remove_filter( 'posts_groupby', array( $instance, 'sorting_groupby_post_id' ) );
+		remove_filter( 'posts_orderby', array( $instance, 'venue_sorting_orderby' ) );
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Tests handle_venue_sorting early return when orderby is not 'venue'.
+	 *
+	 * @covers ::handle_venue_sorting
+	 * @return void
+	 */
+	public function test_venue_sorting_early_return_wrong_orderby(): void {
+		global $wp_the_query;
+
+		// Create a WP_Query for non-venue sorting.
+		$query = new WP_Query(
+			array(
+				'post_type' => Event::POST_TYPE,
+				'orderby'   => 'title',
+			)
+		);
+
+		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited -- Necessary for testing query modifications.
+		$wp_the_query = $query;
+
+		set_current_screen( 'edit-gatherpress_event' );
+
+		$instance = Event_Query::get_instance();
+		$instance->handle_venue_sorting( $query );
+
+		// Should not set venue_sort_order since orderby is not 'venue'.
+		$this->assertEmpty(
+			$query->get( 'venue_sort_order' ),
+			'Should not set venue_sort_order for non-venue orderby'
+		);
+
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Coverage for sortable_columns method.
+	 *
+	 * @covers ::sortable_columns
+	 *
+	 * @return void
+	 */
+	public function test_sortable_columns(): void {
+		$instance = Event_Query::get_instance();
+		$default  = array( 'unit' => 'test' );
+		$expects  = array(
+			'unit'     => 'test',
+			'datetime' => 'datetime',
+			'venue'    => 'venue',
+			'rsvps'    => 'rsvps',
+		);
+
+		$this->assertSame(
+			$expects,
+			$instance->sortable_columns( $default ),
+			'Failed to assert correct sortable columns.'
+		);
+	}
+
+	/**
+	 * Coverage for default_sort method when on the wrong screen.
+	 *
+	 * @covers ::default_sort
+	 *
+	 * @return void
+	 */
+	public function test_default_sort_wrong_screen(): void {
+		$instance = Event_Query::get_instance();
+
+		// Set current screen to a non-event screen.
+		set_current_screen( 'edit-post' );
+
+		// Ensure $_GET is clean.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['orderby'], $_GET['order'] );
+
+		$instance->default_sort();
+
+		// Should return early without modifying $_GET.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		$this->assertArrayNotHasKey( 'orderby', $_GET, 'Should not set orderby on wrong screen.' );
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		$this->assertArrayNotHasKey( 'order', $_GET, 'Should not set order on wrong screen.' );
+
+		// Clean up.
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Coverage for default_sort method when orderby is already set.
+	 *
+	 * @covers ::default_sort
+	 *
+	 * @return void
+	 */
+	public function test_default_sort_orderby_already_set(): void {
+		$instance = Event_Query::get_instance();
+
+		// Set current screen to event edit screen.
+		set_current_screen( 'edit-gatherpress_event' );
+
+		// Set an existing orderby value.
+		$_GET['orderby'] = 'title'; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+
+		$instance->default_sort();
+
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput
+		$this->assertSame( 'title', $_GET['orderby'], 'Should not override existing orderby.' );
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		$this->assertArrayNotHasKey( 'order', $_GET, 'Should not set order when orderby already exists.' );
+
+		// Clean up.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['orderby'], $_GET['order'] );
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Coverage for default_sort method when on the correct screen with no orderby.
+	 *
+	 * @covers ::default_sort
+	 *
+	 * @return void
+	 */
+	public function test_default_sort_sets_defaults(): void {
+		$instance = Event_Query::get_instance();
+
+		// Set current screen to event edit screen.
+		set_current_screen( 'edit-gatherpress_event' );
+
+		// Ensure $_GET is clean.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['orderby'], $_GET['order'] );
+
+		$instance->default_sort();
+
+		// Should set default orderby and order.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput
+		$this->assertSame( 'datetime', $_GET['orderby'], 'Should set orderby to datetime.' );
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput
+		$this->assertSame( 'asc', $_GET['order'], 'Should set order to asc.' );
+
+		// Clean up.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['orderby'], $_GET['order'] );
+		set_current_screen( 'front' );
+	}
+
+	/**
+	 * Coverage for query_vars method.
+	 *
+	 * @covers ::query_vars
+	 *
+	 * @return void
+	 */
+	public function test_query_vars(): void {
+		$instance = Event_Query::get_instance();
+
+		$result = $instance->query_vars( array( 'existing_var' ) );
+
+		$this->assertContains(
+			'gatherpress_event_query',
+			$result,
+			'Should add gatherpress_event_query to query vars.'
+		);
+		$this->assertContains(
+			'existing_var',
+			$result,
+			'Should preserve existing query vars.'
+		);
+		$this->assertCount( 2, $result, 'Should have exactly 2 query vars.' );
+	}
+
+	/**
+	 * Coverage for query_vars method with empty input.
+	 *
+	 * @covers ::query_vars
+	 *
+	 * @return void
+	 */
+	public function test_query_vars_empty_input(): void {
+		$instance = Event_Query::get_instance();
+
+		$result = $instance->query_vars( array() );
+
+		$this->assertContains(
+			'gatherpress_event_query',
+			$result,
+			'Should add gatherpress_event_query even with empty input.'
+		);
+		$this->assertCount( 1, $result, 'Should have exactly 1 query var.' );
+	}
+
+	/**
+	 * Coverage for get_event_counts method with no events.
+	 *
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_get_event_counts_no_events(): void {
+		$instance = Event_Query::get_instance();
+
+		// Reset cached counts and invoke the protected method.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		$counts = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+
+		$this->assertArrayHasKey( 'upcoming', $counts, 'Should have upcoming key.' );
+		$this->assertArrayHasKey( 'past', $counts, 'Should have past key.' );
+		$this->assertSame( 0, $counts['upcoming'], 'Should have 0 upcoming events.' );
+		$this->assertSame( 0, $counts['past'], 'Should have 0 past events.' );
+	}
+
+	/**
+	 * Coverage for get_event_counts method with upcoming events.
+	 *
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_get_event_counts_with_upcoming(): void {
+		$instance = Event_Query::get_instance();
+
+		// Create a published event in the future.
+		$post_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get()->ID;
+
+		$event = new Event( $post_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day +2 hours' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Reset cached counts before querying.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		$counts = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+
+		$this->assertSame( 1, $counts['upcoming'], 'Should have 1 upcoming event.' );
+		$this->assertSame( 0, $counts['past'], 'Should have 0 past events.' );
+	}
+
+	/**
+	 * Coverage for get_event_counts method with past events.
+	 *
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_get_event_counts_with_past(): void {
+		$instance = Event_Query::get_instance();
+
+		// Create a published event in the past.
+		$post_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get()->ID;
+
+		$event = new Event( $post_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '-2 days' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '-2 days +2 hours' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Reset cached counts before querying.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		$counts = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+
+		$this->assertSame( 0, $counts['upcoming'], 'Should have 0 upcoming events.' );
+		$this->assertSame( 1, $counts['past'], 'Should have 1 past event.' );
+	}
+
+	/**
+	 * Coverage for get_event_counts method with mixed events.
+	 *
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_get_event_counts_with_mixed(): void {
+		$instance = Event_Query::get_instance();
+
+		// Create a published upcoming event.
+		$upcoming_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get()->ID;
+
+		$event = new Event( $upcoming_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day +2 hours' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Create a published past event.
+		$past_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get()->ID;
+
+		$event = new Event( $past_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '-2 days' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '-2 days +2 hours' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Create a draft event (should be counted, only trash/auto-draft excluded).
+		$draft_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'draft',
+			)
+		)->get()->ID;
+
+		$event = new Event( $draft_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '+3 days' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '+3 days +2 hours' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Reset cached counts before querying.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		$counts = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+
+		$this->assertSame( 2, $counts['upcoming'], 'Should have 2 upcoming events (draft included).' );
+		$this->assertSame( 1, $counts['past'], 'Should have 1 past event.' );
+	}
+
+	/**
+	 * Coverage for get_event_counts method with a currently running event.
+	 *
+	 * A running event (started but not ended) should count as upcoming
+	 * because datetime_end_gmt >= now, and also as past because
+	 * datetime_start_gmt < now.
+	 *
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_get_event_counts_running_event(): void {
+		$instance = Event_Query::get_instance();
+
+		// Create a published event that is currently running.
+		$post_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get()->ID;
+
+		$event = new Event( $post_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '-1 hour' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '+1 hour' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Reset cached counts before querying.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		$counts = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+
+		// Running events appear in both counts (same logic as Event_Query with inclusive=true).
+		$this->assertSame( 1, $counts['upcoming'], 'Running event should count as upcoming.' );
+		$this->assertSame( 1, $counts['past'], 'Running event should count as past.' );
+	}
+
+	/**
+	 * Coverage for get_event_counts method with events that have no date set.
+	 *
+	 * Events without a date/time set have no row in the gatherpress_events table.
+	 * These should be counted as upcoming, not past.
+	 *
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_get_event_counts_with_no_date(): void {
+		$instance = Event_Query::get_instance();
+
+		// Reset cached counts.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		// Create an event without setting any dates.
+		$this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get();
+
+		$counts = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+
+		$this->assertSame( 1, $counts['upcoming'], 'Event without date should count as upcoming.' );
+		$this->assertSame( 0, $counts['past'], 'Event without date should not count as past.' );
+	}
+
+	/**
+	 * Coverage for get_event_counts caching to class property.
+	 *
+	 * Verifies that repeated calls return the cached result without re-querying.
+	 *
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_get_event_counts_caches_result(): void {
+		$instance = Event_Query::get_instance();
+
+		// Reset cached counts.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		// First call should query the database and cache.
+		$counts = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+		$this->assertSame( 0, $counts['upcoming'], 'Should have 0 upcoming events.' );
+
+		// Verify the property is now cached.
+		$cached = Utility::set_and_get_hidden_property( $instance, 'event_counts', $counts );
+		$this->assertNotNull( $cached, 'Property should be cached after first call.' );
+		$this->assertSame( $counts, $cached, 'Cached value should match returned value.' );
+
+		// Create an event after the first call.
+		$post_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get()->ID;
+
+		$event = new Event( $post_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day +2 hours' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Second call should return cached result (still 0).
+		$counts_cached = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+		$this->assertSame( 0, $counts_cached['upcoming'], 'Cached call should still return 0.' );
+
+		// After resetting the cache, should reflect the new event.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+		$counts_fresh = Utility::invoke_hidden_method( $instance, 'get_event_counts' );
+		$this->assertSame( 1, $counts_fresh['upcoming'], 'Fresh call should return 1 after cache reset.' );
+	}
+
+	/**
+	 * Coverage for views_edit method with no events.
+	 *
+	 * @covers ::views_edit
+	 *
+	 * @return void
+	 */
+	public function test_views_edit_adds_links(): void {
+		$instance = Event_Query::get_instance();
+
+		$view_links = array(
+			'all'     => '<a href="#">All</a>',
+			'publish' => '<a href="#">Published</a>',
+			'draft'   => '<a href="#">Draft</a>',
+		);
+
+		$result = $instance->views_edit( $view_links );
+
+		// Should have original links plus upcoming and past.
+		$this->assertArrayHasKey( 'upcoming', $result, 'Should have upcoming link.' );
+		$this->assertArrayHasKey( 'past', $result, 'Should have past link.' );
+		$this->assertArrayHasKey( 'all', $result, 'Should preserve all link.' );
+		$this->assertArrayHasKey( 'publish', $result, 'Should preserve publish link.' );
+		$this->assertArrayHasKey( 'draft', $result, 'Should preserve draft link.' );
+
+		// Verify count spans are present.
+		$this->assertStringContainsString(
+			'<span class="count">',
+			$result['upcoming'],
+			'Upcoming link should contain count span.'
+		);
+		$this->assertStringContainsString(
+			'<span class="count">',
+			$result['past'],
+			'Past link should contain count span.'
+		);
+
+		// Verify links contain proper query args.
+		$this->assertStringContainsString(
+			'gatherpress_event_query=upcoming',
+			$result['upcoming'],
+			'Upcoming link should have correct query arg.'
+		);
+		$this->assertStringContainsString(
+			'gatherpress_event_query=past',
+			$result['past'],
+			'Past link should have correct query arg.'
+		);
+
+		// Verify sort order: upcoming should be asc, past should be desc.
+		$this->assertStringContainsString(
+			'order=asc',
+			$result['upcoming'],
+			'Upcoming link should sort ascending.'
+		);
+		$this->assertStringContainsString(
+			'order=desc',
+			$result['past'],
+			'Past link should sort descending.'
+		);
+
+		// Verify the labels.
+		$this->assertStringContainsString( 'Upcoming', $result['upcoming'], 'Should contain Upcoming label.' );
+		$this->assertStringContainsString( 'Past', $result['past'], 'Should contain Past label.' );
+	}
+
+	/**
+	 * Coverage for views_edit method link placement.
+	 *
+	 * Upcoming and Past should be inserted after the first link (All).
+	 *
+	 * @covers ::views_edit
+	 *
+	 * @return void
+	 */
+	public function test_views_edit_placement(): void {
+		$instance = Event_Query::get_instance();
+
+		$view_links = array(
+			'all'     => '<a href="#">All</a>',
+			'publish' => '<a href="#">Published</a>',
+		);
+
+		$result = $instance->views_edit( $view_links );
+		$keys   = array_keys( $result );
+
+		$this->assertEquals( 'all', $keys[0], 'First link should be all.' );
+		$this->assertEquals( 'upcoming', $keys[1], 'Second link should be upcoming.' );
+		$this->assertEquals( 'past', $keys[2], 'Third link should be past.' );
+		$this->assertEquals( 'publish', $keys[3], 'Fourth link should be publish.' );
+	}
+
+	/**
+	 * Coverage for views_edit method with active upcoming view.
+	 *
+	 * @covers ::views_edit
+	 *
+	 * @return void
+	 */
+	public function test_views_edit_active_upcoming(): void {
+		$instance = Event_Query::get_instance();
+
+		// Simulate an active upcoming view via GET parameter.
+		$_GET['gatherpress_event_query'] = 'upcoming'; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+
+		// WordPress marks "All" as current by default.
+		$view_links = array(
+			'all' => '<a href="#" class="current" aria-current="page">All</a>',
+		);
+
+		$result = $instance->views_edit( $view_links );
+
+		$this->assertStringContainsString(
+			'class="current"',
+			$result['upcoming'],
+			'Active upcoming link should have current class.'
+		);
+		$this->assertStringContainsString(
+			'aria-current="page"',
+			$result['upcoming'],
+			'Active upcoming link should have aria-current attribute.'
+		);
+
+		// Past should not be marked as current.
+		$this->assertStringNotContainsString(
+			'class="current"',
+			$result['past'],
+			'Past link should not have current class.'
+		);
+
+		// "All" should have its current class removed.
+		$this->assertStringNotContainsString(
+			'class="current"',
+			$result['all'],
+			'All link should not have current class when filter is active.'
+		);
+
+		// Clean up.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['gatherpress_event_query'] );
+	}
+
+	/**
+	 * Coverage for views_edit method with active past view.
+	 *
+	 * @covers ::views_edit
+	 *
+	 * @return void
+	 */
+	public function test_views_edit_active_past(): void {
+		$instance = Event_Query::get_instance();
+
+		// Simulate an active past view via GET parameter.
+		$_GET['gatherpress_event_query'] = 'past'; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+
+		$view_links = array(
+			'all' => '<a href="#">All</a>',
+		);
+
+		$result = $instance->views_edit( $view_links );
+
+		$this->assertStringContainsString(
+			'class="current"',
+			$result['past'],
+			'Active past link should have current class.'
+		);
+
+		// Upcoming should not be marked as current.
+		$this->assertStringNotContainsString(
+			'class="current"',
+			$result['upcoming'],
+			'Upcoming link should not have current class.'
+		);
+
+		// Clean up.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['gatherpress_event_query'] );
+	}
+
+	/**
+	 * Coverage for views_edit method with no active event query filter.
+	 *
+	 * @covers ::views_edit
+	 *
+	 * @return void
+	 */
+	public function test_views_edit_no_active_filter(): void {
+		$instance = Event_Query::get_instance();
+
+		// Ensure no event query filter is set.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['gatherpress_event_query'] );
+
+		// WordPress marks "All" as current by default.
+		$view_links = array(
+			'all' => '<a href="#" class="current" aria-current="page">All</a>',
+		);
+
+		$result = $instance->views_edit( $view_links );
+
+		// Neither Upcoming nor Past should be marked as current.
+		$this->assertStringNotContainsString(
+			'class="current"',
+			$result['upcoming'],
+			'Upcoming link should not have current class without filter.'
+		);
+		$this->assertStringNotContainsString(
+			'class="current"',
+			$result['past'],
+			'Past link should not have current class without filter.'
+		);
+
+		// "All" should keep its current class.
+		$this->assertStringContainsString(
+			'class="current"',
+			$result['all'],
+			'All link should keep current class when no filter is active.'
+		);
+	}
+
+	/**
+	 * Coverage for views_edit adding current class to "All" when WordPress omits it.
+	 *
+	 * When default_sort() adds orderby/order to $_GET, WordPress's
+	 * is_base_request() returns false and omits the current class from "All".
+	 *
+	 * @covers ::views_edit
+	 *
+	 * @return void
+	 */
+	public function test_views_edit_all_gets_current_when_missing(): void {
+		$instance = Event_Query::get_instance();
+
+		// Ensure no event query filter is set.
+		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		unset( $_GET['gatherpress_event_query'] );
+
+		// Simulate WordPress not adding current class due to default_sort()
+		// adding extra $_GET params that break is_base_request().
+		$view_links = array(
+			'all' => '<a href="#">All</a>',
+		);
+
+		$result = $instance->views_edit( $view_links );
+
+		// "All" should get the current class added by views_edit.
+		$this->assertStringContainsString(
+			'class="current"',
+			$result['all'],
+			'All link should get current class when no filter is active.'
+		);
+		$this->assertStringContainsString(
+			'aria-current="page"',
+			$result['all'],
+			'All link should get aria-current when no filter is active.'
+		);
+	}
+
+	/**
+	 * Coverage for views_edit method with event counts displayed.
+	 *
+	 * @covers ::views_edit
+	 * @covers ::get_event_counts
+	 *
+	 * @return void
+	 */
+	public function test_views_edit_displays_counts(): void {
+		$instance = Event_Query::get_instance();
+
+		// Create a published upcoming event.
+		$post_id = $this->mock->post(
+			array(
+				'post_type'   => Event::POST_TYPE,
+				'post_status' => 'publish',
+			)
+		)->get()->ID;
+
+		$event = new Event( $post_id );
+		$event->save_datetimes(
+			array(
+				'datetime_start' => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day' ) ),
+				'datetime_end'   => gmdate( 'Y-m-d H:i:s', strtotime( '+1 day +2 hours' ) ),
+				'timezone'       => 'UTC',
+			)
+		);
+
+		// Reset cached counts after event creation so fresh queries run.
+		Utility::set_and_get_hidden_property( $instance, 'event_counts', null );
+
+		$view_links = array(
+			'all' => '<a href="#">All</a>',
+		);
+
+		$result = $instance->views_edit( $view_links );
+
+		// Upcoming should show count of 1.
+		$this->assertStringContainsString(
+			'<span class="count">(1)</span>',
+			$result['upcoming'],
+			'Upcoming link should show count of 1.'
+		);
+
+		// Past should show count of 0.
+		$this->assertStringContainsString(
+			'<span class="count">(0)</span>',
+			$result['past'],
+			'Past link should show count of 0.'
 		);
 	}
 }
